feat: add admin page, fix connections lack problem
By adding admin page, it was needed to make all pages work well with
admin page, as well as 'admin' role was added to application. So, it was
needed to add authorization which was done. About
problem with connection: sometimes it happend there were no available
connection becasuse of them not being returned to connection pool, so it
was fixed. About refactoring: repeated code was removed.